+++ src/Adapter/Alias/CommandHandler/AddAlias.php
@@ -24,25 +24,38 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-namespace PrestaShop\PrestaShop\Core\Domain\Alias\CommandHandler;
+declare(strict_types=1);
 
+namespace PrestaShop\PrestaShop\Adapter\Alias\CommandHandler;
+
+use PrestaShop\PrestaShop\Adapter\Alias\Repository\AliasRepository;
 use PrestaShop\PrestaShop\Core\Domain\Alias\Command\AddAliasCommand;
-use PrestaShop\PrestaShop\Core\Domain\Alias\Exception\InvalidAliasIdException;
-use PrestaShop\PrestaShop\Core\Domain\Alias\ValueObject\AliasId;
-use PrestaShop\PrestaShop\Core\Exception\CoreException;
+use PrestaShop\PrestaShop\Core\Domain\Alias\CommandHandler\AddAliasCommandInterface;
 
-/**
- * Interface for services that handle command which adds new alias
- */
-interface AddAliasCommandHandlerInterface
+class AddAlias implements AddAliasCommandInterface
 {
     /**
-     * @param AddAliasCommand $command
-     *
-     * @return AliasId[]
-     *
-     * @throws InvalidAliasIdException
-     * @throws CoreException
+     * @var AliasRepository
      */
-    public function handle(AddAliasCommand $command): array;
+    private $aliasRepository;
+
+    /**
+     * @param AliasRepository $aliasRepository
+     */
+    public function __construct(
+        AliasRepository $aliasRepository
+    ) {
+        $this->aliasRepository = $aliasRepository;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function handle(AddAliasCommand $command): array
+    {
+        return $this->aliasRepository->create(
+            $command->getSearch(),
+            $command->getAliases()
+        );
+    }
 }
